requirements.txt
cffi==1.15.0
click==8.0.4
cryptography==36.0.2
Deprecated==1.2.13
Flask==2.0.3
gunicorn==20.1.0
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.1
packaging==21.3
pycparser==2.21
pycryptodome==3.14.1
pyparsing==3.0.7
redis==4.1.4
Werkzeug==2.0.3
wrapt==1.14.0
